Create checout trait
